Add Stats method
It returns stats one time for bulk of containers
Old method for streaming is removed.
