+++ supervisor/event.go
@@ -21,7 +21,8 @@ const (
 	UpdateContainerEventType  EventType = "updateContainer"
 	CreateCheckpointEventType EventType = "createCheckpoint"
 	DeleteCheckpointEventType EventType = "deleteCheckpoint"
-	StatsEventType            EventType = "events"
+	StatsEventType            EventType = "stats"
+	SubscribeStatsEventType   EventType = "subscribeStats"
 	UnsubscribeStatsEventType EventType = "unsubscribeStats"
 	StopStatsEventType        EventType = "stopStats"
 	OOMEventType              EventType = "oom"
@@ -57,7 +58,8 @@ type Event struct {
 	Checkpoint    *runtime.Checkpoint
 	Err           chan error
 	StartResponse chan StartResponse
-	Stats         chan interface{}
+	StatsStream   chan interface{}
+	Stats         *runtime.Stat
 }
 
 type Handler interface {
