+++ supervisor/stats.go
@@ -4,37 +4,15 @@ type StatsEvent struct {
 	s *Supervisor
 }
 
-type UnsubscribeStatsEvent struct {
-	s *Supervisor
-}
-
-type StopStatsEvent struct {
-	s *Supervisor
-}
-
 func (h *StatsEvent) Handle(e *Event) error {
 	i, ok := h.s.containers[e.ID]
 	if !ok {
 		return ErrContainerNotFound
 	}
-	e.Stats = h.s.statsCollector.collect(i.container)
-	return nil
-}
-
-func (h *UnsubscribeStatsEvent) Handle(e *Event) error {
-	i, ok := h.s.containers[e.ID]
-	if !ok {
-		return ErrContainerNotFound
+	st, err := i.container.Stats()
+	if err != nil {
+		return err
 	}
-	h.s.statsCollector.unsubscribe(i.container, e.Stats)
-	return nil
-}
-
-func (h *StopStatsEvent) Handle(e *Event) error {
-	i, ok := h.s.containers[e.ID]
-	if !ok {
-		return ErrContainerNotFound
-	}
-	h.s.statsCollector.stopCollection(i.container)
+	e.Stats = st
 	return nil
 }
